+++ src/iranges/__init__.py
@@ -15,6 +15,6 @@ except PackageNotFoundError:  # pragma: no cover
 finally:
     del version, PackageNotFoundError
 
-from .IRanges import IRanges
+from .iranges import IRanges
 from .utils import normalize_array
 from .irangeslist import CompressedIRangesList