+++ src/iranges/irangeslist.py
@@ -1,10 +1,12 @@
-from typing import List, Optional, Sequence, Union
+from __future__ import annotations
+
+from typing import Any, Dict, List, Optional, Sequence, Union
 
 import biocutils as ut
 from compressed_lists import CompressedList, Partitioning
 from compressed_lists.split_generic import _generic_register_helper, splitAsCompressedList
 
-from .IRanges import IRanges, _combine_IRanges
+from .iranges import IRanges, _combine_IRanges
 
 __author__ = "Jayaram Kancherla"
 __copyright__ = "Jayaram Kancherla"
@@ -19,7 +21,7 @@ class CompressedIRangesList(CompressedList):
         unlist_data: IRanges,
         partitioning: Partitioning,
         element_metadata: Optional[dict] = None,
-        metadata: Optional[dict] = None,
+        metadata: Optional[Union[Dict[str, Any], ut.NamedList]] = None,
         **kwargs,
     ):
         """Initialize a CompressedIRangesList.
@@ -52,8 +54,8 @@ class CompressedIRangesList(CompressedList):
         cls,
         lst: List[IRanges],
         names: Optional[Union[ut.Names, Sequence[str]]] = None,
-        metadata: Optional[dict] = None,
-    ) -> "CompressedIRangesList":
+        metadata: Optional[Union[Dict[str, Any], ut.NamedList]] = None,
+    ) -> CompressedIRangesList:
         """Create a `CompressedIRangesList` from a regular list.
 
         This concatenates the list of `IRanges` objects.